Move public-related classes to publicpackage
