+++ app/src/main/res/navigation/navigation_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/publicFragment"
-        android:name="com.shapeapp.shape.ui.PublicFragment"
+        android:name="com.shapeapp.shape.ui.publicpackage.PublicFragment"
         android:label="fragment_public"
         tools:layout="@layout/fragment_public">
         <action
